Extract page generation from app.js
